requirements.txt
# Desktop App Dependencies
PySide6>=6.4.0
pandas>=1.3.0
requests>=2.28.0
img2pdf==0.6.3
Pillow>=9.0.0
numpy<2.0.0
onnxruntime>=1.14.0
